Create HR-Diagram(star classification)
calculating the mass of the milky way using the Gaia catalog of stars 50pc away, finding their density, and assuming the milky way is cylindrical. Produces a massive underestimate by a factor of 1000. Does not take into account the milky ways SMBH or dark matter or possible variations in density as you move through the milky way.